Fix casting issue with long ids
This commit resolves issue #116. The issue occurs when a long ID is read as a float. This means that
the indexes are converted to standard form, resulting in no merges when compared to the benchmarks. to resolve this
dtype={0:str} has been added to every pd.read_csv() where elara reads intermediate data. This forces the index to
be read as a str.

Resolves #116